Improve code structure
