+++ lib/notifications.dart
@@ -1,1 +1,2 @@
-export 'src/notification_service.dart';
+export 'src/notifications/notification_service.dart';
+export 'src/notifications/player_control_command.dart';
