+++ lib/src/notifications/player_control_command.dart
@@ -1,0 +1,4 @@
+enum PlayerControlCommand {
+  NEXT_TRACK,
+  PREVIOUS_TRACK,
+}
